prod minor fix

+++ build-scripts/build-prod.sh
@@ -1,4 +1,14 @@
-#!/bin/bash
+#!/usr/bin/env bash
+set -e
+
+if [ -f ./configs/.production.env ]; then
+  set -o allexport
+  source ./configs/.production.env
+  set +o allexport
+fi
+
 tsc
 mkdir -p ./dist/configs/
-cp ./configs/.production.env ./dist/configs/
+if [ -f ./configs/.production.env ]; then
+  cp ./configs/.production.env ./dist/configs/
+fi
